requirements.txt
Flask==2.0.1
Flask_WTF==0.15.1
Pillow==8.4.0
Werkzeug==2.0.1
WTForms==2.3.3
gunicorn==20.1.0
